Remove dbg
