--- contracts/exec-callee/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "exec-callee"
-version = "0.1.0"
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[features]
-native-simulator = ["ckb-std/native-simulator"]
-
-[dependencies]
-ckb-std = { path = "../../" }
