--- contracts/exec-callee/src/lib.rs
@@ -1,6 +1,0 @@
-#![cfg_attr(not(feature = "native-simulator"), no_std)]
-
-#[cfg(feature = "native-simulator")]
-pub mod entry;
-#[cfg(feature = "native-simulator")]
-pub mod error;
