+++ test/simulator/src/exec_callee.rs
@@ -1,7 +1,18 @@
 extern crate alloc;
 
-#[path = "../../../contracts/exec-callee/src/entry.rs"]
-mod entry;
+mod entry {
+    use core::result::Result;
+    use crate::error::Error;
+
+    pub fn main() -> Result<(), Error> {
+        let argv = ckb_std::env::argv();
+        ckb_std::debug!("argv: {:?}", argv);
+        assert_eq!(argv.len(), 2);
+        assert_eq!(argv[0].to_bytes(), b"Hello World");
+        assert_eq!(argv[1].to_bytes(), "你好".as_bytes());
+        Ok(())
+    }
+}
 
 pub mod error {
     use ckb_std::error::SysError;
